// src/api/stetsom/model/assetVersion.ts
/**
 * Generated by orval v8.10.0 🍺
 * Do not edit manually.
 * Stetsom API
 * API documentation for Stetsom
 * OpenAPI spec version: 1.0.0
 */

export interface AssetVersion {
  version_id: string;
  file_url: string;
  size_bytes: number;
  width?: number;
  height?: number;
  created_at: string;
  created_by: string;
}